Add files via upload
Transaction fee reward calculation converted to day wise

+++ contracts_Transcation/Timing.sol
@@ -6,16 +6,16 @@ import "./Declaration.sol";
 
 abstract contract Timing is Declaration {
 
-    function currentLPDay() public view returns (uint64) {
+    function currentLPDay() public view returns (uint256) {
         return _getNow() >= LAUNCH_TIME ? _currentLPDay() : 0;
     } 
 
-    function _currentLPDay() internal view returns (uint64) {
+    function _currentLPDay() internal view returns (uint256) {
         return _LPDayFromStamp(_getNow());
     }
 
-    function _LPDayFromStamp(uint256 _timestamp) internal view returns (uint64) {
-        return uint64((_timestamp - LAUNCH_TIME) / SECONDS_IN_DAY_LP);
+    function _LPDayFromStamp(uint256 _timestamp) internal view returns (uint256) {
+        return uint256((_timestamp - LAUNCH_TIME) / SECONDS_IN_DAY_LP);
     }
 
     function currentGriseWeek() public view returns (uint256) {
@@ -23,24 +23,24 @@ abstract contract Timing is Declaration {
         return((curGriseDay % 7) > 0)? (curGriseDay / 7) + 1 : (curGriseDay / 7);
     }
 
-    function currentGriseDay() public view returns (uint64) {
+    function currentGriseDay() public view returns (uint256) {
         return _getNow() >= LAUNCH_TIME ? _currentGriseDay() : 0;
     }
 
-    function _currentGriseDay() internal view returns (uint64) {
+    function _currentGriseDay() internal view returns (uint256) {
         return _griseDayFromStamp(_getNow());
     }
 
-    function _nextGriseDay() internal view returns (uint64) {
+    function _nextGriseDay() internal view returns (uint256) {
         return _currentGriseDay() + 1;
     }
 
-    function _previousGriseDay() internal view returns (uint64) {
+    function _previousGriseDay() internal view returns (uint256) {
         return _currentGriseDay() - 1;
     }
 
-    function _griseDayFromStamp(uint256 _timestamp) internal view returns (uint64) {
-        return uint64((_timestamp - LAUNCH_TIME) / SECONDS_IN_DAY);
+    function _griseDayFromStamp(uint256 _timestamp) internal view returns (uint256) {
+        return uint256((_timestamp - LAUNCH_TIME) / SECONDS_IN_DAY);
     }
 
     function _getNow() internal view returns (uint256) {
